Wrapped the old ServerPropertyDatabase into SerializedShardDatabase
