+++ Source/ACE.Database/SerializedShardDatabase.cs
@@ -7,6 +7,9 @@ using System.Threading.Tasks;
 using ACE.Database.Entity;
 using ACE.Database.Models.Shard;
 using ACE.Entity.Enum;
+using Microsoft.EntityFrameworkCore;
+using System.Linq;
+using log4net;
 
 namespace ACE.Database
 {
@@ -18,9 +21,12 @@ namespace ACE.Database
 
         private Thread _workerThread;
 
+        public ShardPropertyTables Config;
+
         internal SerializedShardDatabase(ShardDatabase shardDatabase)
         {
             _wrappedDatabase = shardDatabase;
+            Config = new ShardPropertyTables();
         }
 
         public void Start()
@@ -394,6 +400,191 @@ namespace ACE.Database
         public void SetCharacterAccessLevelByName(string name, AccessLevel accessLevel, Action<uint> callback)
         {
             throw new NotImplementedException();
+        }
+    }
+
+    public class ShardPropertyTables
+    {
+        private static readonly ILog log = LogManager.GetLogger(System.Reflection.MethodBase.GetCurrentMethod().DeclaringType);
+
+        public void AddBool(string key, bool value, string description = "")
+        {
+            var stat = new ConfigPropertiesBoolean
+            {
+                Key = key,
+                Value = value,
+                Description = description
+            };
+
+            using (var context = new ShardDbContext())
+            {
+                context.ConfigPropertiesBoolean.Add(stat);
+
+                context.SaveChanges();
+            }
+        }
+
+        public void ModifyBool(ConfigPropertiesBoolean stat)
+        {
+            using (var context = new ShardDbContext())
+            {
+                context.Entry(stat).State = EntityState.Modified;
+
+                context.SaveChanges();
+            }
+        }
+
+        public ConfigPropertiesBoolean GetBool(string key)
+        {
+            using (var context = new ShardDbContext())
+                return context.ConfigPropertiesBoolean.AsNoTracking().FirstOrDefault(r => r.Key == key);
+        }
+
+        public bool BoolExists(string key)
+        {
+            using (var context = new ShardDbContext())
+                return context.ConfigPropertiesBoolean.Any(r => r.Key == key);
+        }
+
+        public void AddLong(string key, long value, string description = "")
+        {
+            var stat = new ConfigPropertiesLong
+            {
+                Key = key,
+                Value = value,
+                Description = description
+            };
+
+            using (var context = new ShardDbContext())
+            {
+                context.ConfigPropertiesLong.Add(stat);
+
+                context.SaveChanges();
+            }
+        }
+
+        public void ModifyLong(ConfigPropertiesLong stat)
+        {
+            using (var context = new ShardDbContext())
+            {
+                context.Entry(stat).State = EntityState.Modified;
+
+                context.SaveChanges();
+            }
+        }
+
+        public ConfigPropertiesLong GetLong(string key)
+        {
+            using (var context = new ShardDbContext())
+                return context.ConfigPropertiesLong.AsNoTracking().FirstOrDefault(r => r.Key == key);
+        }
+
+        public bool LongExists(string key)
+        {
+            using (var context = new ShardDbContext())
+                return context.ConfigPropertiesLong.Any(r => r.Key == key);
+        }
+
+        public void AddDouble(string key, double value, string description = "")
+        {
+            var stat = new ConfigPropertiesDouble
+            {
+                Key = key,
+                Value = value,
+                Description = description
+            };
+
+            using (var context = new ShardDbContext())
+            {
+                context.ConfigPropertiesDouble.Add(stat);
+
+                context.SaveChanges();
+            }
+        }
+
+        public void ModifyDouble(ConfigPropertiesDouble stat)
+        {
+            using (var context = new ShardDbContext())
+            {
+                context.Entry(stat).State = EntityState.Modified;
+
+                context.SaveChanges();
+            }
+        }
+
+        public ConfigPropertiesDouble GetDouble(string key)
+        {
+            using (var context = new ShardDbContext())
+                return context.ConfigPropertiesDouble.AsNoTracking().FirstOrDefault(r => r.Key == key);
+        }
+
+        public bool DoubleExists(string key)
+        {
+            using (var context = new ShardDbContext())
+                return context.ConfigPropertiesDouble.Any(r => r.Key == key);
+        }
+
+        public void AddString(string key, string value, string description = "")
+        {
+            var stat = new ConfigPropertiesString
+            {
+                Key = key,
+                Value = value,
+                Description = description
+            };
+
+            using (var context = new ShardDbContext())
+            {
+                context.ConfigPropertiesString.Add(stat);
+
+                context.SaveChanges();
+            }
+        }
+
+        public void ModifyString(ConfigPropertiesString stat)
+        {
+            using (var context = new ShardDbContext())
+            {
+                context.Entry(stat).State = EntityState.Modified;
+
+                context.SaveChanges();
+            }
+        }
+
+        public ConfigPropertiesString GetString(string key)
+        {
+            using (var context = new ShardDbContext())
+                return context.ConfigPropertiesString.AsNoTracking().FirstOrDefault(r => r.Key == key);
+        }
+
+        public bool StringExists(string key)
+        {
+            using (var context = new ShardDbContext())
+                return context.ConfigPropertiesString.Any(r => r.Key == key);
+        }
+
+        public List<ConfigPropertiesBoolean> GetAllBools()
+        {
+            using (var context = new ShardDbContext())
+                return context.ConfigPropertiesBoolean.AsNoTracking().ToList();
+        }
+
+        public List<ConfigPropertiesLong> GetAllLongs()
+        {
+            using (var context = new ShardDbContext())
+                return context.ConfigPropertiesLong.AsNoTracking().ToList();
+        }
+
+        public List<ConfigPropertiesDouble> GetAllDoubles()
+        {
+            using (var context = new ShardDbContext())
+                return context.ConfigPropertiesDouble.AsNoTracking().ToList();
+        }
+
+        public List<ConfigPropertiesString> GetAllStrings()
+        {
+            using (var context = new ShardDbContext())
+                return context.ConfigPropertiesString.AsNoTracking().ToList();
         }
     }
 }
